fix warning

+++ src/bwapi/BWAPILIB/Unitset.cpp
@@ -79,7 +79,7 @@ namespace BWAPI
   }
   void Unitset::setClientInfo(int clientInfo, int index) const
   {
-    this->setClientInfo((void*)clientInfo, index);
+    this->setClientInfo((void*)(long long int)clientInfo, index);
   }
 
   Unitset Unitset::getUnitsInRadius(int radius, const UnitFilter &pred) const
